+++ src/Common/IEnvironment.cs
@@ -1,6 +1,7 @@
-namespace Common;
+﻿namespace Common;
 
 public interface IEnvironment
 {
     public string GetFolderPath(Environment.SpecialFolder folder);
+    string GetFolderPath(Environment.SpecialFolder folder, Environment.SpecialFolderOption option);
 }