+++ src/Recyclarr.Tests/Command/Initialization/Init/InitializeAppDataPathTest.cs
@@ -1,4 +1,5 @@
 using AutoFixture.NUnit3;
+using Common;
 using NSubstitute;
 using NUnit.Framework;
 using Recyclarr.Command;
@@ -13,25 +14,29 @@ namespace Recyclarr.Tests.Command.Initialization.Init;
 public class InitializeAppDataPathTest
 {
     [Test, AutoMockData]
-    public void Do_not_override_path_if_null(
+    public void Use_default_app_data_if_not_specified(
+        [Frozen] IEnvironment env,
         [Frozen] IAppPaths paths,
         SonarrCommand cmd,
         InitializeAppDataPath sut)
     {
+        env.GetFolderPath(Arg.Any<Environment.SpecialFolder>(), Arg.Any<Environment.SpecialFolderOption>())
+            .Returns("app_data");
+
         sut.Initialize(cmd);
 
-        paths.DidNotReceiveWithAnyArgs().SetAppDataPath(default!);
+        env.Received().GetFolderPath(Environment.SpecialFolder.ApplicationData, Environment.SpecialFolderOption.Create);
+        paths.Received().SetAppDataPath("app_data");
     }
 
     [Test, AutoMockData]
-    public void Override_path_if_not_null(
+    public void Use_specified_app_data_if_user_provided(
         [Frozen] IAppPaths paths,
         SonarrCommand cmd,
         InitializeAppDataPath sut)
     {
         cmd.AppDataDirectory = "path";
         sut.Initialize(cmd);
-
         paths.Received().SetAppDataPath("path");
     }
 }